Current universe conversions pulled into the bijection module
