+++ cosmos-model/src/main/scala/com/mesosphere/universe/v3/model/PackagingVersion.scala
@@ -8,15 +8,19 @@ case object V3PackagingVersion extends PackagingVersion("3.0")
 
 object PackagingVersion {
 
-  val allVersions: Map[String, PackagingVersion] = {
-    Seq(V2PackagingVersion, V3PackagingVersion).map(v => v.show -> v).toMap
+  val allVersions = Seq(V2PackagingVersion, V3PackagingVersion)
+
+  private[this] val allVersionsString = allVersions.map(_.show).mkString(", ")
+
+  private[this] val allVersionsIndex: Map[String, PackagingVersion] = {
+    allVersions.map(v => v.show -> v).toMap
   }
 
   def apply(s: String): Try[PackagingVersion] = {
-    allVersions.get(s) match {
+    allVersionsIndex.get(s) match {
       case Some(v) => Return(v)
       case _ => Throw(new IllegalArgumentException(
-        s"Expected one of [${allVersions.keySet.mkString(", ")}] for packaging version, but found [$s]"
+        s"Expected one of [$allVersionsString] for packaging version, but found [$s]"
       ))
     }
   }
