--- cosmos-server/src/main/scala/com/mesosphere/cosmos/converter/ConversionFailure.scala
@@ -1,3 +1,0 @@
-package com.mesosphere.cosmos.converter
-
-final case class ConversionFailure(message: String) extends RuntimeException(message)
