+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/converter/Label.scala
@@ -1,11 +1,10 @@
 package com.mesosphere.cosmos.converter
 
-import com.mesosphere.cosmos.converter.Common._
-import com.mesosphere.cosmos.converter.Universe._
 import com.mesosphere.cosmos.internal
 import com.mesosphere.cosmos.label
 import com.mesosphere.cosmos.rpc
 import com.mesosphere.universe
+import com.mesosphere.universe.bijection.UniverseConversions._
 import com.twitter.bijection.Conversion.asMethod
 import com.twitter.bijection.{Bijection, Conversion}
