+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/handler/ListVersionsHandler.scala
@@ -1,11 +1,10 @@
 package com.mesosphere.cosmos.handler
 
-import com.mesosphere.cosmos.converter.Common._
-import com.mesosphere.cosmos.converter.Universe._
 import com.mesosphere.cosmos.http.RequestSession
 import com.mesosphere.cosmos.repository.PackageCollection
 import com.mesosphere.cosmos.rpc
 import com.mesosphere.universe
+import com.mesosphere.universe.bijection.UniverseConversions._
 import com.twitter.bijection.Conversion.asMethod
 import com.twitter.util.Future
