+++ cosmos-server/src/test/scala/com/mesosphere/cosmos/SelectedPackageSpec.scala
@@ -1,13 +1,13 @@
 package com.mesosphere.cosmos
 
 import cats.data.Xor
-import com.mesosphere.cosmos.converter.Universe._
 import com.mesosphere.cosmos.handler.PackageSearchHandler
 import com.mesosphere.cosmos.repository.PackageCollection
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
 import com.mesosphere.cosmos.rpc.v1.circe.Encoders._
 import com.mesosphere.cosmos.rpc.v1.model.{SearchRequest, SearchResponse, SearchResult}
 import com.mesosphere.universe
+import com.mesosphere.universe.bijection.UniverseConversions._
 import com.mesosphere.universe.v2.circe.Decoders._
 import com.mesosphere.universe.v2.circe.Encoders._
 import com.mesosphere.universe.v2.model._
